Day 20, faster approach
Visited the solution megathread on r/adventofcode. Picked up the
approximation for the lower bound for part 1 and the faster algorithm
for part 2.

Now part 2 is even faster than previously with the non proven hack. The
only question remaining is, if the part 2 answer could ever be lower
than part 1. This solution made that assumption with no proof.